% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{changes}
\alias{changes}
\title{View changes notes.}
\usage{
changes(pkg = "diezeit")
}
\arguments{
\item{pkg}{Set to the default "diezeit". Other packages make no sense.}
}
\description{
\code{changes} brings up the NEWS file of the package.
}
\examples{
\dontrun{
changes()
}
}

