% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{zeit_content}
\alias{zeit_content}
\title{Search for content}
\usage{
zeit_content(query, fields = "all", limit = 10, offset = 0)
}
\arguments{
\item{query}{the main search query; single character value or character vector.}

\item{fields}{partially select output fields; set to \code{"all"} by default.}

\item{limit}{limit the amount of matches to return; set to 10 by default.}

\item{offset}{offset for the list of matches; set to 0 by default.}
}
\value{
a list of articles that matches to the query.
}
\description{
\code{zeit_content} exposes an unfiltered search for articles.
You can set search queries, paginate, sort and partially select the fields,
that should be returned. Articles, that match your query, are returned with
a reduced set of meta data.
}
\examples{
\dontrun{
zeit_content("")
}
}

