% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{zeit_search}
\alias{zeit_search}
\title{Search the ZEIT archive}
\usage{
zeit_search(endpoint, query, fields = "all", limit = 10, offset = 0,
  print = TRUE)
}
\arguments{
\item{endpoint}{one of \code{author}, \code{content}, \code{department}, \code{keyword},
\code{product} or \code{series} -- provides specific search functionalities.}

\item{query}{the main search query; single character value or character vector.}

\item{fields}{partially select output fields; set to \code{"all"} by default.}

\item{limit}{limit the amount of matches to return; set to 10 by default.}

\item{offset}{offset for the list of matches; set to 0 by default.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
a list of matches to the query.
}
\description{
\code{zeit_search} exposes a search for ZEIT archive items.
You can set search queries, paginate, sort and partially select the fields,
that should be returned. Articles, that match your query, are returned with
a reduced set of meta data.
}
\examples{
\dontrun{
zeit_search("")
}
}

