% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diezeit.R
\docType{package}
\name{diezeit}
\alias{diezeit}
\alias{diezeit-package}
\title{ZEIT ONLINE Content API}
\description{
ZEIT ONLINE Content API
}
\seealso{
\code{\link{zeit_client}} for client information and usage,
\code{\link{zeit_content}} for unfiltered article search.
}

