% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{zeit_client}
\alias{zeit_client}
\title{Observe your usage}
\usage{
zeit_client(print = TRUE)
}
\arguments{
\item{print}{if \code{TRUE} (default), the client information is printed.}
}
\value{
a list of information about the client and API usage
}
\description{
\code{zeit_client} does not provide content per se,
but lets you get information about your API usage.
}
\examples{
\dontrun{
zeit_client()
}
}

