% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{zeit_get}
\alias{zeit_get}
\title{Get content from the ZEIT archive}
\source{
\url{http://developer.zeit.de/docs/}
}
\usage{
zeit_get(endpoint, id, fields, limit = 10, offset = 0, sort, print = TRUE)
}
\arguments{
\item{endpoint}{one of \code{author}, \code{content}, \code{department}, \code{keyword},
\code{product} or \code{series} -- see \code{\link{zeit_search}}.}

\item{id}{item id.}

\item{fields}{partially select output fields, as string value or vector of strings for multiple fields.}

\item{limit}{limit the amount of matches to return; set to \code{10} by default.}

\item{offset}{offset for the list of matches; set to \code{0} by default.}

\item{sort}{sort search results by any of the returned \code{fields}.
Vector of two (\code{c([field], [direction])}), giving field and direction keyword.
Direction keywords are \code{asc} and \code{desc} for an ascending or descending
sort order respectively. Multiple sort orders are accepted as \code{list} of such vectors.}

\item{print}{if \code{TRUE} (default) the search results are printed.}
}
\value{
Article content or list of articles.
}
\description{
\code{zeit_get} will get you all available data for an article,
or provide a pre-filtered search for all articles.
}
\details{
\emph{Endpoints}

The API is structured into several endpoints that provide specific functionalities:
\tabular{lll}{
\tab author \tab content by this author \cr
\tab content \tab get content by ID \cr
\tab department \tab content from this department \cr
\tab keyword \tab content about this keyword \cr
\tab product \tab content from this product \cr
\tab series \tab content in this series
}
}
\examples{
\dontrun{
zeit_get("")
}
}

